requirements.txt
absl-py==2.1.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.13.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.5
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
deepface==0.0.93
distro==1.9.0
dlib-bin==19.24.6
dnspython==2.6.1
easyocr==1.7.1
entrypoints==0.4
executing==2.0.1
face_recognition_models==0.3.0
fastapi==0.115.6
filelock==3.15.4
fire==0.7.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==1.34.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-vision==3.7.4
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.65.1
grpcio-status==1.48.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.7
imageio==2.34.2
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgspec==0.19.0
mtcnn==1.0.0
mysql-connector-python==9.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
openai==1.51.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
Pympler==1.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
setuptools==72.1.0
shapely==2.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.11.0
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
streamlit==1.22.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.7.24
tokenizers==0.20.0
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
validators==0.34.0
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
# wikipedia==1.4.0
wrapt==1.16.0
xgboost==2.1.1
zipp==3.20.2
